Migrate from Lottie to Compottie for KMP support
This commit replaces the Android-specific Lottie library with Compottie, a Kotlin Multiplatform alternative.

Key changes:
- Replaced `com.airbnb.android:lottie-compose` with `io.github.alexzhirkevich:compottie`.
- Migrated `YralLottieAnimation`, `YralRemoteLottieAnimation`, and `PreloadLottieAnimation` from expect/actual implementations to a common `commonMain` implementation using Compottie.
- Removed all platform-specific Lottie files (`.android.kt`, `.ios.kt`).
- Updated `GameIconBubbles` to use `ContentScale.FillBounds` for better animation scaling.
- Added a new constructor to `YralException` to include a cause.
