+++ gradle/libs.versions.toml
@@ -42,7 +42,6 @@ kotlinx-datetime = "0.7.1"
 kotlinxSerialization = "1.9.0"
 ksp = "2.1.21-2.0.2"
 ktor = "3.3.0"
-lottie = "6.6.6"
 media3Exoplayer = "1.6.1"
 mixpanel = "8.2.0"
 mixpanel-session-replay = "1.0.0"
@@ -60,6 +59,7 @@ branch="5.19.0"
 play-services-ads-identifier="18.0.1"
 emoji2 = "1.5.0"
 window = "1.5.0"
+compottie = "2.0.1"
 
 [libraries]
 
@@ -98,6 +98,12 @@ compose-ui-tooling-preview = { module = "androidx.compose.ui:ui-tooling-preview"
 compose-foundation = { module = "androidx.compose.foundation:foundation", version.ref = "compose" }
 compose-material3 = { module = "androidx.compose.material3:material3", version.ref = "compose-material3" }
 compose-ui-backhandler = { module = "org.jetbrains.compose.ui:ui-backhandler", version.ref = "composeMultiplatform" }
+
+compottie = { module = "io.github.alexzhirkevich:compottie", version.ref = "compottie" }
+compottie-lite = { module = "io.github.alexzhirkevich:compottie-lite", version.ref = "compottie" }
+compottie-dot = { module = "io.github.alexzhirkevich:compottie-dot", version.ref = "compottie" }
+compottie-network = { module = "io.github.alexzhirkevich:compottie-network", version.ref = "compottie" }
+compottie-resources = { module = "io.github.alexzhirkevich:compottie-resources", version.ref = "compottie" }
 
 decompose-decompose = { group = "com.arkivanov.decompose", name = "decompose", version.ref = "decompose" }
 decompose-extensions-compose = { group = "com.arkivanov.decompose", name = "extensions-compose", version.ref = "decompose" }
@@ -156,8 +162,6 @@ ktor-http = { module = "io.ktor:ktor-http", version.ref = "ktor" }
 ktor-client-logging = { module = "io.ktor:ktor-client-logging", version.ref = "ktor" }
 ktor-client-okhttp = { module = "io.ktor:ktor-client-okhttp", version.ref = "ktor" }
 ktor-serialization-kotlinx-json = { module = "io.ktor:ktor-serialization-kotlinx-json", version.ref = "ktor" }
-
-lottie = { module = "com.airbnb.android:lottie-compose", version.ref = "lottie" }
 
 mixpanel-android = { module = "com.mixpanel.android:mixpanel-android", version.ref = "mixpanel" }
 mixpanel-session-replay-android = { module = "com.mixpanel.android:mixpanel-android-session-replay", version.ref = "mixpanel-session-replay" }
